Project/FastAPI and Docker practice/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
transformers==4.30.2
pydantic==2.5.2
redis==5.0.1 
python-dotenv
